Rename TicketTemplate to ITILTemplate
Migrate database schema
