+++ front/itiltemplatepredefinedfield.form.php
@@ -30,7 +30,27 @@
  * ---------------------------------------------------------------------
  */
 
+use Glpi\Event;
+
 include ('../inc/includes.php');
 
-$dropdown = new TicketTemplate();
-include (GLPI_ROOT . "/front/dropdown.common.form.php");
+Session ::checkRight('itiltemplate', UPDATE);
+
+$item = new ITILTemplatePredefinedField();
+
+// Use masiveaction system to manage add value
+if (isset($_POST["massiveaction"])) {
+   $item->check(-1, UPDATE, $_POST);
+   if (isset($_POST['items_tickets_id']) && isset($_POST['add_items_id'])) {
+      $_POST['items_tickets_id'] = $_POST['items_tickets_id']."_".$_POST['add_items_id'];
+   }
+   if ($item->add($_POST)) {
+      Event::log($_POST["itiltemplates_id"], "itiltemplate", 4, "maintain",
+                 //TRANS: %s is the user login
+                 sprintf(__('%s adds predefined field'), $_SESSION["glpiname"]));
+   }
+   Html::back();
+
+}
+
+Html::displayErrorAndDie("lost");
