+++ tests/web/APIRest.php
@@ -343,8 +343,8 @@ class APIRest extends APIBaseClass {
     */
    public function testGetItemsCommonDBChild() {
       // test the case have DBChild not have entities_id
-      $ticketTemplate = new \TicketTemplate();
-      $ticketTMF = new \TicketTemplateMandatoryField();
+      $ticketTemplate = new \ITILTemplate();
+      $ticketTMF = new \ITILTemplateMandatoryField();
 
       $tt_id = $ticketTemplate->add([
          'entities_id' => 0,
@@ -353,15 +353,15 @@ class APIRest extends APIBaseClass {
       $this->boolean((bool)$tt_id)->isTrue();
 
       $ttmf_id = $ticketTMF->add([
-         'tickettemplates_id' => $tt_id,
+         'itiltemplates_id' => $tt_id,
          'num'                => 7
       ]);
       $this->boolean((bool)$ttmf_id)->isTrue();
 
       $data = $this->query('getItems',
                            ['query'     => [
-                               'searchText' => ['tickettemplates_id' => "^".$tt_id."$"]],
-                            'itemtype'   => 'TicketTemplateMandatoryField',
+                               'searchText' => ['itiltemplates_id' => "^".$tt_id."$"]],
+                            'itemtype'   => 'ITILTemplateMandatoryField',
                             'headers'    => ['Session-Token' => $this->session_token]],
                            200);
       if (isset($data['headers'])) {
